added batch class

+++ lamp/lamp_data/src/main/java/com/lamp/ApplicantQuizResponse.java
@@ -1,8 +1,8 @@
 package com.lamp;
 
 public class ApplicantQuizResponse {
-	private int[] questionId;
-	private int[] response;
+	private byte[] questionId;
+	private byte[] response;
